Trait azuki_syntax::visitor::AstVisitor

pub trait AstVisitor {
    type LExprResult;
    type ExprResult;
    type TyResult;
    type StmtResult;
    type ProgramResult;
    type FuncResult;
    fn visit_program(&mut self, program: &Program) -> Self::ProgramResult { ... }

    fn visit_func(&mut self, func: &FuncStmt) -> Self::FuncResult { ... }

    fn visit_func_param(&mut self, _param: &FuncParam) -> Self::StmtResult { ... }

    fn visit_ty(&mut self, _ty: &TyDef) -> Self::TyResult { ... }

    fn visit_expr(&mut self, expr: &Expr) -> Self::ExprResult { ... }

    fn visit_literal_expr(&mut self, _expr: &LiteralExpr) -> Self::ExprResult { ... }

    fn visit_ident_expr(&mut self, _expr: &Ident) -> Self::ExprResult { ... }

    fn visit_assign_expr(&mut self, expr: &AssignExpr) -> Self::ExprResult { ... }

    fn visit_lexpr(&mut self, _expr: &Expr) -> Self::LExprResult { ... }

    fn visit_binary_expr(&mut self, expr: &BinaryExpr) -> Self::ExprResult { ... }

    fn visit_unary_expr(&mut self, expr: &UnaryExpr) -> Self::ExprResult { ... }

    fn visit_call_expr(&mut self, expr: &CallExpr) -> Self::ExprResult { ... }

    fn visit_as_expr(&mut self, expr: &AsExpr) -> Self::ExprResult { ... }

    fn visit_stmt(&mut self, stmt: &Stmt) -> Self::StmtResult { ... }

    fn visit_block_stmt(&mut self, stmt: &BlockStmt) -> Self::StmtResult { ... }

    fn visit_while_stmt(&mut self, stmt: &WhileStmt) -> Self::StmtResult { ... }

    fn visit_if_stmt(&mut self, stmt: &IfStmt) -> Self::StmtResult { ... }

    fn visit_expr_stmt(&mut self, stmt: &Expr) -> Self::StmtResult { ... }

    fn visit_decl_stmt(&mut self, stmt: &DeclStmt) -> Self::StmtResult { ... }

    fn visit_return_stmt(&mut self, stmt: &ReturnStmt) -> Self::StmtResult { ... }

    fn visit_break_stmt(&mut self, _span: Span) -> Self::StmtResult { ... }

    fn visit_continue_stmt(&mut self, _span: Span) -> Self::StmtResult { ... }

    fn visit_empty_stmt(&mut self, _span: Span) -> Self::StmtResult { ... }
}

A handy trait for traversing the abstract syntax tree.

Every function inside this visitor should call the corresponding walk_<type> function in order to traverse deeper into the tree.

Associated Types

type LExprResult

type ExprResult

type TyResult

type StmtResult

type ProgramResult

type FuncResult

Loading content...
Provided methods

fn visit_program(&mut self, program: &Program) -> Self::ProgramResult

fn visit_func(&mut self, func: &FuncStmt) -> Self::FuncResult

fn visit_func_param(&mut self, _param: &FuncParam) -> Self::StmtResult

fn visit_ty(&mut self, _ty: &TyDef) -> Self::TyResult

fn visit_expr(&mut self, expr: &Expr) -> Self::ExprResult

fn visit_literal_expr(&mut self, _expr: &LiteralExpr) -> Self::ExprResult

fn visit_ident_expr(&mut self, _expr: &Ident) -> Self::ExprResult

fn visit_assign_expr(&mut self, expr: &AssignExpr) -> Self::ExprResult

fn visit_lexpr(&mut self, _expr: &Expr) -> Self::LExprResult

fn visit_binary_expr(&mut self, expr: &BinaryExpr) -> Self::ExprResult

fn visit_unary_expr(&mut self, expr: &UnaryExpr) -> Self::ExprResult

fn visit_call_expr(&mut self, expr: &CallExpr) -> Self::ExprResult

fn visit_as_expr(&mut self, expr: &AsExpr) -> Self::ExprResult

fn visit_stmt(&mut self, stmt: &Stmt) -> Self::StmtResult

fn visit_block_stmt(&mut self, stmt: &BlockStmt) -> Self::StmtResult

fn visit_while_stmt(&mut self, stmt: &WhileStmt) -> Self::StmtResult

fn visit_if_stmt(&mut self, stmt: &IfStmt) -> Self::StmtResult

fn visit_expr_stmt(&mut self, stmt: &Expr) -> Self::StmtResult

fn visit_decl_stmt(&mut self, stmt: &DeclStmt) -> Self::StmtResult

fn visit_return_stmt(&mut self, stmt: &ReturnStmt) -> Self::StmtResult

fn visit_break_stmt(&mut self, _span: Span) -> Self::StmtResult

fn visit_continue_stmt(&mut self, _span: Span) -> Self::StmtResult

fn visit_empty_stmt(&mut self, _span: Span) -> Self::StmtResult

Loading content...